+++ src/constant/index.ts
@@ -1,0 +1,8 @@
+export const constants = {
+    default_lang: "en",
+    default_theme: "conan",
+    default_font: "Cambria, Cochin, Georgia, Times, 'Times New Roman', serif",
+    default_position: "tr",
+    default_short_cut: "Alt+l",
+    default_auto_start: "true",
+}
