+++ src/locals/resources.ts
@@ -1,0 +1,57 @@
+export const resources = {
+    en: {
+        "home": {
+            "search_text": "Type a snippet name to search...",
+            "snippet_not_find": "No snippet found."
+        },
+        "share": {
+            "import": "Import",
+            "window_title": "Share",
+            "export_message": "The exported data has been copied to the clipboard!!"
+        },
+        "setting": {
+            "font_family": "Systerm font",
+            "language": "Language",
+            "theme": "Theme",
+            "short_cut": "ShortClasut",
+            "auto_start": "Auto start",
+            "position": "Position",
+            "recommendation": "After modifying the settings, the application needs to be restarted to take effect!!!",
+        },
+        "themes_name": {
+            "conan": "Conan",
+            "yuanshen": "Genshin",
+            "qiyuanzui": "Seven deadly sins",
+            "lol": "League of Legends",
+            "pure": "Pure color"
+        }
+    },
+    zh: {
+        "home": {
+            "search_text": "输入片段名称进行搜索...",
+            "snippet_not_find": "未找到片段。"
+        },
+        "share": {
+            "import": "导入",
+            "window_title": "分享",
+            "export_message": "导出的数据已复制到剪贴板！"
+        },
+        "setting": {
+            "font_family": "系统字体",
+            "language": "语言",
+            "theme": "主题",
+            "short_cut": "快捷键",
+            "auto_start": "开机自启",
+            "position": "位置",
+            "recommendation": "修改设置后需要重启应用才能生效！！！",
+        },
+        "themes_name": {
+            "conan": "柯南",
+            "yuanshen": "原神",
+            "qiyuanzui": "七原罪",
+            "lol": "英雄联盟",
+            "pure": "纯色"
+        }
+    }
+}
+
