+++ src/pages/Settings/index.tsx
@@ -7,8 +7,10 @@ import { Short_cut } from "@/pages/Settings/items/Short_cut";
 import { Theme } from "@/pages/Settings/items/Theme";
 import { Auto_start } from "@/pages/Settings/items/Auto_start";
 import { Position } from "@/pages/Settings/items/Position";
+import { useTranslation } from "react-i18next";
 export const Settings = () => {
     const navigate = useNavigate();
+    const { t } = useTranslation();
     function handle_return() {
         navigate('/');
     }
@@ -23,6 +25,7 @@ export const Settings = () => {
                 <Auto_start />
                 <Position />
                 <Theme />
+                <p className="mt-2 text-xs text-red-600">{t("setting.recommendation")}</p>
             </div>
         </div>
