+++ src/pages/Settings/items/Auto_start.tsx
@@ -2,15 +2,15 @@ import { Setting_item } from "@/pages/Settings/components/Setting_item"
 import { Switch } from "@/components/ui/switch"
 import { useTranslation } from "react-i18next"
 import { useEffect, useState } from "react"
-import LocalStorage from '@/lib/localstorage'
+import { Localstorage } from '@/lib/localstorage'
 import { useAutoStart } from "@/hooks/use_auto_start"
+import { constants } from "@/constant"
 
 export const Auto_start = () => {
     const { t } = useTranslation()
-    const [auto_start, set_auto_start] = useState<boolean>(LocalStorage.getItem("auto_start") === "true")
+    const [auto_start, set_auto_start] = useState<boolean>(Localstorage.getItemWithDefault("auto_start", constants.default_auto_start) === "true")
     useEffect(() => {
-        LocalStorage.setItem("auto_start", auto_start.toString())
-        useAutoStart()
+        Localstorage.runFnWithLocalStorage("auto_start", auto_start.toString(), () => useAutoStart(auto_start))
     }, [auto_start])
     return <Setting_item title={t('setting.auto_start') + ":"}>
         <div className="flex items-center space-x-2">
