+++ src/pages/Share/JsonEditor.tsx
@@ -9,7 +9,7 @@ interface Props {
 }
 
 export const JsonEditor: FC<Props> = ({ value, setValue }) => {
-    const { themes, local } = useContext(ThemeContext)
+    const { themes, theme: local } = useContext(ThemeContext)
 
     return <AceEditor
         mode="json"
